Refactor parse multi hd keys
