+++ include/URRegistryFFI/lib_ur_registry_ffi.h
@@ -7,8 +7,14 @@ struct ExternError {
 
 void keystone_sdk_destroy_string(const char* cstring);
 
-// Wallet
+// Sync
+const char* parse_crypto_hd_key(struct ExternError*, const char* ur_type, const char* cbor_hex);
+const char* parse_crypto_account(struct ExternError*, const char* ur_type, const char* cbor_hex);
 const char* parse_crypto_multi_accounts(struct ExternError*, const char* ur_type, const char* cbor_hex);
+
+// Utils
+const char* get_uncompressed_key(struct ExternError*, const char* compressed_key)
+const char* derive_public_key(struct ExternError*, const char* xpub, const char* path)
 
 // BTC
 const char* generate_crypto_psbt(struct ExternError*, const char* psbt_hex);
