+++ libs/ur-registry-ffi/src/lib.rs
@@ -4,13 +4,14 @@ pub mod bitcoin;
 pub mod cosmos;
 pub mod ethereum;
 mod export;
-pub mod extend;
+pub mod sync;
 pub mod keystone;
 pub mod near;
 pub mod solana;
 pub mod tron;
 pub mod cardano;
-mod util;
+mod util_internal;
+mod utils;
 pub mod sui;
 
 ffi_support::define_string_destructor!(keystone_sdk_destroy_string);
