+++ libs/ur-registry-ffi/src/sync/crypto_account.rs
@@ -1,0 +1,74 @@
+use crate::export;
+use anyhow::format_err;
+use anyhow::Error;
+use hex;
+use serde_json::json;
+use ur_registry::crypto_account::CryptoAccount;
+use ur_registry::crypto_output::CryptoOutput;
+use ur_registry::registry_types::{CRYPTO_ACCOUNT};
+use ur_registry::traits::From;
+use crate::sync::crypto_multi_accounts::MultiAccounts;
+
+impl Into<MultiAccounts> for CryptoAccount {
+    fn into(self) -> MultiAccounts {
+        let hd_keys = self
+            .get_output_descriptors()
+            .iter()
+            .map(|output| CryptoOutput::get_hd_key(output).unwrap_or_default())
+            .map(|crypto_hd_key| (&crypto_hd_key).into())
+            .collect();
+        MultiAccounts {
+            master_fingerprint: hex::encode(self.get_master_fingerprint()),
+            keys: hd_keys,
+            device: None,
+        }
+    }
+}
+
+export! {
+    @Java_com_keystone_sdk_KeystoneNativeSDK_parseCryptoAccount
+    fn parse_crypto_account(ur_type: &str, cbor_hex: &str) -> String {
+        if CRYPTO_ACCOUNT.get_type() != ur_type {
+            return json!({"error": "type not match"}).to_string();
+        }
+
+        let parse_signature = || -> Result<MultiAccounts, Error> {
+            let cbor = hex::decode(cbor_hex.to_string())?;
+            let crypto_account = CryptoAccount::from_cbor(cbor).map_err(|_| format_err!(""))?;
+            let multi_accounts = crypto_account.into();
+            Ok(multi_accounts)
+        };
+        match parse_signature() {
+            Ok(multi_accounts) => json!(multi_accounts).to_string(),
+            Err(_) => json!({"error": "crypto account is invalid"}).to_string(),
+        }
+    }
+
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn test_parse_crypto_account() {
+        let hd_keys_cbor = "A2011A52006EA0028AD9012FA502F403582102FEF03A2BD3DE113F1DC1CDB1E69AA4D935DC3458D542D796F5827ABBB1A58B5E06D90130A3018A182CF5183CF500F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F4035821033F1EDDF1D1BB2762FCFA67FBC35E12DC9968CD2587ADA055210E84F780C1109A06D90130A3018A182CF5183CF501F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102C5FCF766AD77A0C254834D57CE3E6120A2BE5C266E9BABE8A047D1A53CB34F9E06D90130A3018A182CF5183CF502F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102CD0B648CF944CBA7E6BE97BF1F17F0EAB7B9E600D181C421B3BCE6E7F6D941F006D90130A3018A182CF5183CF503F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F40358210351F72104E737E94C7CC66E33307C74D5BBF19216800157AD34EBFE232F23C75106D90130A3018A182CF5183CF504F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102037F8C5FC1074E654FF11619A8BF28DCC3DB5D037191F08EB5722252AF57A4A606D90130A3018A182CF5183CF505F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582103A441895DFBE9C7B3BF8EBA0CE461465A14350D902DF163A0B3F06E4F4843E54F06D90130A3018A182CF5183CF506F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102A8DCDF480733A5B7FB331C9464B7E0EDF5206D8581FE3E26BFD6DE38C8063D4C06D90130A3018A182CF5183CF507F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F4035821037A1A6A48B09D4E3A01223B37C9D1212D8DA20746302009956168E1EA3BD3E0C806D90130A3018A182CF5183CF508F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582103C6F04A813F23799940B6FA44C6CA48ABE04DE9FBB8133B7342DBABC95B0EA48106D90130A3018A182CF5183CF509F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665";
+        let expect_result = "{\"device\":null,\"keys\":[{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/0'/0/0\",\"public_key\":\"02fef03a2bd3de113f1dc1cdb1e69aa4d935dc3458d542d796f5827abbb1a58b5e\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/1'/0/0\",\"public_key\":\"033f1eddf1d1bb2762fcfa67fbc35e12dc9968cd2587ada055210e84f780c1109a\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/2'/0/0\",\"public_key\":\"02c5fcf766ad77a0c254834d57ce3e6120a2be5c266e9babe8a047d1a53cb34f9e\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/3'/0/0\",\"public_key\":\"02cd0b648cf944cba7e6be97bf1f17f0eab7b9e600d181c421b3bce6e7f6d941f0\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/4'/0/0\",\"public_key\":\"0351f72104e737e94c7cc66e33307c74d5bbf19216800157ad34ebfe232f23c751\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/5'/0/0\",\"public_key\":\"02037f8c5fc1074e654ff11619a8bf28dcc3db5d037191f08eb5722252af57a4a6\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/6'/0/0\",\"public_key\":\"03a441895dfbe9c7b3bf8eba0ce461465a14350d902df163a0b3f06e4f4843e54f\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/7'/0/0\",\"public_key\":\"02a8dcdf480733a5b7fb331c9464b7e0edf5206d8581fe3e26bfd6de38c8063d4c\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/8'/0/0\",\"public_key\":\"037a1a6a48b09d4e3a01223b37c9d1212d8da20746302009956168e1ea3bd3e0c8\",\"xfp\":\"52006ea0\"},{\"chain\":\"ETH\",\"chain_code\":\"\",\"extended_public_key\":\"\",\"name\":\"Keystone\",\"note\":\"account.ledger_live\",\"path\":\"m/44'/60'/9'/0/0\",\"public_key\":\"03c6f04a813f23799940b6fa44c6ca48abe04de9fbb8133b7342dbabc95b0ea481\",\"xfp\":\"52006ea0\"}],\"master_fingerprint\":\"52006ea0\"}";
+
+        assert_eq!(
+            expect_result,
+            parse_crypto_account("crypto-account", hd_keys_cbor)
+        );
+    }
+
+    #[test]
+    fn test_parse_crypto_account_wrong_type() {
+        let hd_key_cbor = "A2011A52006EA0028AD9012FA502F403582102FEF03A2BD3DE113F1DC1CDB1E69AA4D935DC3458D542D796F5827ABBB1A58B5E06D90130A3018A182CF5183CF500F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F4035821033F1EDDF1D1BB2762FCFA67FBC35E12DC9968CD2587ADA055210E84F780C1109A06D90130A3018A182CF5183CF501F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102C5FCF766AD77A0C254834D57CE3E6120A2BE5C266E9BABE8A047D1A53CB34F9E06D90130A3018A182CF5183CF502F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102CD0B648CF944CBA7E6BE97BF1F17F0EAB7B9E600D181C421B3BCE6E7F6D941F006D90130A3018A182CF5183CF503F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F40358210351F72104E737E94C7CC66E33307C74D5BBF19216800157AD34EBFE232F23C75106D90130A3018A182CF5183CF504F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102037F8C5FC1074E654FF11619A8BF28DCC3DB5D037191F08EB5722252AF57A4A606D90130A3018A182CF5183CF505F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582103A441895DFBE9C7B3BF8EBA0CE461465A14350D902DF163A0B3F06E4F4843E54F06D90130A3018A182CF5183CF506F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582102A8DCDF480733A5B7FB331C9464B7E0EDF5206D8581FE3E26BFD6DE38C8063D4C06D90130A3018A182CF5183CF507F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F4035821037A1A6A48B09D4E3A01223B37C9D1212D8DA20746302009956168E1EA3BD3E0C806D90130A3018A182CF5183CF508F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665D9012FA502F403582103C6F04A813F23799940B6FA44C6CA48ABE04DE9FBB8133B7342DBABC95B0EA48106D90130A3018A182CF5183CF509F500F400F4021A52006EA0030509684B657973746F6E650A736163636F756E742E6C65646765725F6C697665";
+        let expect_result = "{\"error\":\"type not match\"}";
+
+        assert_eq!(
+            expect_result,
+            parse_crypto_account("crypto-hdkey", hd_key_cbor)
+        );
+    }
+}
