+++ libs/ur-registry-ffi/src/sync/mod.rs
@@ -1,2 +1,3 @@
 pub mod crypto_hd_key;
+pub mod crypto_account;
 pub mod crypto_multi_accounts;
