+++ libs/ur-registry-ffi/src/util_internal/mod.rs
